Add files via upload
It is an android application developed with the implementation of Android Studio where "Java" has been used for the backend programming and "xml" has been used for the UI design. The application allows a user to communicate with another user through Video Call. Here, at the beginning when the user registers his or her number then a token is sent to his or her mobile phone. After receiving the token, the user inputs the token in the " Code Verification edit text attribute" of the application. If the code is correct the user gets registered to "Video Call Elite". Now the current user can search other users, send friend requests to other users,  as well as receive friend requests from other  people. The user can call another user if that user is in the current user's friend list.